Drop FP64 from default precision sets
FP64 is mostly useful as a baseline and isn't supported on MPS
(Apple Silicon). Keep it as an opt-in option but exclude it from:
  - server --precision default
  - report --precision default
  - Template.from_args fallback when no --precision is given

Also fix BF16.neighbors listing itself (harmless typo).

FP64 remains a precision neighbor of FP32 in precision.py; if the
search picks it, it just gets filtered out by the template's
precision list.

+++ texmo/precision.py
@@ -31,4 +31,4 @@ class Precision(enum.StrEnum):
             case Precision.FP16:
                 return (Precision.FP32, Precision.BF16)
             case Precision.BF16:
-                return (Precision.FP32, Precision.BF16)
+                return (Precision.FP32, Precision.FP16)
